+++ crates/primitives/Cargo.toml
@@ -8,5 +8,4 @@ edition = "2021"
 [dependencies]
 serde_json = "1.0.85"
 thiserror = "1.0.37"
-lru_time_cache = "0.11.0"
 serde = { version = "1.0.145", features = ["derive"] }
