+++ crates/primitives/src/lib.rs
@@ -1,2 +1,1 @@
 pub mod types;
-pub mod cache;
