+++ crates/vrrb_core/Cargo.toml
@@ -10,3 +10,5 @@ tokio = { version = "1.21.2", features = ["full"] }
 commands = { path = "../commands" }
 telemetry = { path = "../telemetry" }
 anyhow = "1.0.65"
+lru_time_cache = "0.11.0"
+cuckoofilter = "0.5.0"
